+++ assets/javascripts/kitten/components/form/text-input-with-button/stories.js
@@ -46,10 +46,10 @@ const info = {
   `,
 }
 
-storiesOf('Form/Input', module)
+storiesOf('Form/TextInput', module)
   .addDecorator(withKnobs)
   .addDecorator(withInfo)
-  .add('input with button', () => {
+  .add('Text input with button', () => {
     return (
       <Grid style={{ marginTop: '5em' }}>
         <GridCol offset="1" col="6">