+++ assets/javascripts/kitten/components/form/text-input/stories.js
@@ -50,10 +50,10 @@ const info = {
   `,
 }
 
-storiesOf('Form/Input', module)
+storiesOf('Form/TextInput', module)
   .addDecorator(withKnobs)
   .addDecorator(withInfo)
-  .add('input', () => {
+  .add('Text input', () => {
     return (
       <Grid style={{ marginTop: '5em' }}>
         <GridCol offset="1" col="6">